Fixes for making executables that run.
My text executable crashes *once* it's running, but that's at least
better than not being recognized as a valid executable at all.

There's obviously something I'm still doing wrong when re-building the
data files. What I see when reverse engineering the executable itself is
that the GameMaker archive is being loaded and parsed at runtime. I'm
pretty confident that making a 1-for-1 swap (as far as size is
concerned) should not be a problem. My only conclusion is that I must
not be outputting completely correct offsets.

To test that theory, I used `extract.rb` to re-extract the new archive
I'd created. The MD5 checksums match on all the files. So, either I'm
wrong and that's *not* the problem...or, my parser is simply better
than/making different assumptions than their parser.